infrastructure/functions/divebar_mirror/requirements.txt
# Google Drive API
google-api-python-client>=2.100.0
google-auth>=2.23.0

# BigQuery
google-cloud-bigquery>=3.13.0

# Cloud Functions framework
functions-framework>=3.0.0
